Create Python: Find a motif
Find a motif PLUS 2 flanking bp in a seq: User can search for a specific motif or search for a category of motifs 
(i.e. ALL of the consensus/1bp/2bp off motifs) Position and occurrence will be reported. You can change the consensus motif and seq 
to ANY length/nucleotides. Appliable to amino acids search.